everything works if we do not query outside grid...
