+++ metaconfigure/description.json
@@ -407,6 +407,12 @@
     "src/ACEtk/interpretation/MP1.hpp",
     "src/ACEtk/interpretation/DEDX1/CRTP.hpp",
     "src/ACEtk/interpretation/DEDX1/StandardWithoutCutoff.hpp",
+    "src/ACEtk/interpretation/DEDX1/CRTP/src/densities.hpp",
+    "src/ACEtk/interpretation/DEDX1/CRTP/src/energies.hpp",
+    "src/ACEtk/interpretation/DEDX1/CRTP/src/S0.hpp",
+    "src/ACEtk/interpretation/DEDX1/CRTP/src/S1.hpp",
+    "src/ACEtk/interpretation/DEDX1/CRTP/src/stoppingPowers.hpp",
+    "src/ACEtk/interpretation/DEDX1/CRTP/src/temperatures.hpp",
     "src/ACEtk/interpretation/EL03/AtomicLevels.hpp",
     "src/ACEtk/interpretation/EL03/MottScatteringCorrection.hpp",
     "src/ACEtk/interpretation/EL03/RadiativeStoppingPower.hpp",
